[cli][test][logging] adding the ability to detect any error messages that are logged during execution * By adding a MemoryHandler to the root logger and filtering out all messages that are not of   level 'error', we can store any errors that occur during test runs in a buffer. * The number of errors that occurred will be reported at the end, and the errors themselves will   be logged so they can be addressed.
