+++ stream_alert_cli/logger.py
@@ -14,6 +14,27 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 import logging
+import logging.handlers
+
+
+class SuppressNonErrors(logging.Filter):
+    """Simple logging filter to allow only caching error messages to a MemoryHandler"""
+    def filter(self, record):
+        return record.levelno == logging.ERROR
+
+
+class SuppressNoise(logging.Filter):
+    """Simple logging filter for suppressing specific log messagses that we
+    do not want to print during testing. Add any suppressions to the tuple.
+    """
+
+    def filter(self, record):
+        suppress_starts_with = (
+            'Starting download from S3',
+            'Completed download in'
+        )
+        return not record.getMessage().startswith(suppress_starts_with)
+
 
 LOGGER_SA = logging.getLogger('StreamAlert')
 LOGGER_SA.setLevel(logging.INFO)
@@ -33,3 +54,14 @@ for logger in logging.Logger.manager.loggerDict:
     if logger.startswith('StreamAlert'):
         continue
     logging.getLogger(logger).setLevel(logging.CRITICAL)
+
+# Get a logging MemoryHandler with a default buffer size of 1000
+# We don't care about assigning a target to this handler since these logs
+# will not actually be written out to disk, etc
+LOG_ERROR_HANDLER = logging.handlers.MemoryHandler(1000)
+
+# Add a filter to suppress everything that is not an error
+LOG_ERROR_HANDLER.addFilter(SuppressNonErrors())
+
+# Add the MemoryHandler to the root logger to capture all logs in all loggers
+logging.getLogger().addHandler(LOG_ERROR_HANDLER)
